Create rememberAppThemeController and then use preferences for app launch to be sued for reading thememode state
